pyproject.toml
[tool.poetry]
name = "ylab"
version = "0.1.0"
description = "FastAPI HW"
authors = ["samolin <samolin.vlad@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
sqlalchemy = "^2.0.19"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
pytest = "^7.4.0"
requests = "^2.31.0"
httpx = "^0.24.1"
pydantic-settings = "^2.0.2"
pre-commit = "^3.3.3"
fastapi-cache2 = "^0.2.1"
redis = "^4.6.0"
types-redis = "^4.6.0.3"
asyncpg = "^0.28.0"
pydantic-core = "^2.4.0"
pytest-asyncio = "^0.21.1"
pandas = "^2.0.3"
openpyxl = "^3.1.2"
celery = {extras = ["librabbitmq"], version = "^5.3.1"}
flower = "^2.0.0"
amqp = "^5.1.1"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
asyncio_mode="auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
